+++ test/MoqxRelayTest.cpp
@@ -6,395 +6,10 @@
  * Copyright (c) OpenMOQ contributors.
  */
 
-#include "MoqxRelay.h"
-#include "TestUtils.h"
+#include "MoqxRelayTestFixture.h"
 #include "UpstreamProvider.h"
-#include <folly/coro/BlockingWait.h>
-#include <folly/io/async/EventBase.h>
-#include <folly/portability/GMock.h>
-#include <folly/portability/GTest.h>
-#include <moxygen/MoQTrackProperties.h>
-#include <moxygen/events/MoQFollyExecutorImpl.h>
-#include <moxygen/relay/MoQForwarder.h>
-#include <moxygen/test/MockMoQSession.h>
-#include <moxygen/test/Mocks.h>
-
-using namespace testing;
-using namespace moxygen;
-using namespace openmoq::moqx;
 
 namespace moxygen::test {
-
-using openmoq::moqx::test::makeBuf;
-
-const TrackNamespace kTestNamespace{{"test", "namespace"}};
-const TrackNamespace kAllowedPrefix{{"test"}};
-const FullTrackName kTestTrackName{kTestNamespace, "track1"};
-
-// TestMoQExecutor that can be driven for tests
-class TestMoQExecutor : public MoQFollyExecutorImpl, public folly::DrivableExecutor {
-public:
-  explicit TestMoQExecutor() : MoQFollyExecutorImpl(&evb_) {}
-  ~TestMoQExecutor() override = default;
-
-  void add(folly::Func func) override { MoQFollyExecutorImpl::add(std::move(func)); }
-
-  // Implements DrivableExec::drive
-  void drive() override {
-    // Run the event loop until there is nothing left to do
-    // (simulate a "tick" for test event loop)
-    if (auto* evb = getBackingEventBase()) {
-      evb->loopOnce();
-    }
-  }
-
-private:
-  folly::EventBase evb_;
-};
-
-// Test fixture for MoqxRelay tests
-// This provides a skeleton for testing MoqxRelay functionality.
-// Full integration tests with real sessions will be added when implementing
-// multi-publisher support.
-class MoQRelayTest : public ::testing::Test {
-protected:
-  void SetUp() override {
-    exec_ = std::make_shared<TestMoQExecutor>();
-    relay_ = std::make_shared<MoqxRelay>(config::CacheConfig{.maxCachedTracks = 0});
-    relay_->setAllowedNamespacePrefix(kAllowedPrefix);
-  }
-
-  void TearDown() override { relay_.reset(); }
-
-  // Helper to create a mock session
-  std::shared_ptr<MockMoQSession> createMockSession() {
-    auto session = std::make_shared<NiceMock<MockMoQSession>>(exec_);
-    ON_CALL(*session, getNegotiatedVersion())
-        .WillByDefault(Return(std::optional<uint64_t>(kVersionDraftCurrent)));
-    auto state = getOrCreateMockState(session);
-    return session;
-  }
-
-  // Helper to create a mock subscription handle for publish calls
-  std::shared_ptr<Publisher::SubscriptionHandle> createMockSubscriptionHandle() {
-    SubscribeOk ok;
-    ok.requestID = RequestID(0);
-    ok.trackAlias = TrackAlias(0);
-    ok.expires = std::chrono::milliseconds(0);
-    ok.groupOrder = GroupOrder::Default;
-    auto handle = std::make_shared<NiceMock<MockSubscriptionHandle>>(std::move(ok));
-    return handle;
-  }
-
-  // Helper to remove a session from the relay and clean up mock state
-  void removeSession(std::shared_ptr<MoQSession> sess) { cleanupMockSession(std::move(sess)); }
-
-  // Helper to create a mock consumer with default actions
-  std::shared_ptr<MockTrackConsumer> createMockConsumer() {
-    auto consumer = std::make_shared<NiceMock<MockTrackConsumer>>();
-    ON_CALL(*consumer, setTrackAlias(_))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-    ON_CALL(*consumer, publishDone(_))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-    return consumer;
-  }
-
-  // Helper to subscribe a session to a track
-  std::shared_ptr<SubscriptionHandle> subscribeToTrack(
-      std::shared_ptr<MoQSession> session,
-      const FullTrackName& trackName,
-      std::shared_ptr<TrackConsumer> consumer,
-      RequestID requestID = RequestID(0),
-      bool addToState = true,
-      folly::Optional<SubscribeErrorCode> expectedError = folly::none
-  ) {
-    SubscribeRequest sub;
-    sub.fullTrackName = trackName;
-    sub.requestID = requestID;
-    sub.locType = LocationType::LargestObject;
-    std::shared_ptr<SubscriptionHandle> handle{nullptr};
-    withSessionContext(session, [&]() {
-      auto task = relay_->subscribe(std::move(sub), consumer);
-      auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-      if (expectedError.has_value()) {
-        // Expect an error and, if present, verify the error code matches
-        EXPECT_FALSE(res.hasValue());
-        // res has an error; compare error code when available
-        const auto& err = res.error();
-        // If SubscribeError exposes a code accessor or public member, use it.
-        // Assuming err.code exists or err.errorCode() accessor:
-        // Adjust the line below to match actual error API.
-        EXPECT_EQ(err.errorCode, *expectedError);
-      } else {
-        EXPECT_TRUE(res.hasValue());
-        handle = *res;
-        if (addToState) {
-          getOrCreateMockState(session)->subscribeHandles.push_back(handle);
-        }
-      }
-    });
-    return handle;
-  }
-
-  // Helper to set up RequestContext for a session (simulates incoming request)
-  template <typename Func>
-  auto withSessionContext(std::shared_ptr<MoQSession> session, Func&& func) -> decltype(func()) {
-    folly::RequestContextScopeGuard guard;
-    folly::RequestContext::get()->setContextData(
-        sessionRequestToken(),
-        std::make_unique<MoQSession::MoQSessionRequestData>(std::move(session))
-    );
-    return func();
-  }
-
-  static const folly::RequestToken& sessionRequestToken() {
-    static folly::RequestToken token("moq_session");
-    return token;
-  }
-
-  // Helper to simulate session cleanup for mock sessions
-  // Real MoQRelaySession calls cleanup() which invokes callbacks on stored
-  // state. For mock sessions, we need to manually track and clean up.
-  struct MockSessionState {
-    std::shared_ptr<MoQSession> session;
-    std::vector<std::shared_ptr<TrackConsumer>> publishConsumers;
-    // Handles for results from publishNamespace, subscribeNamespace,
-    // and subscribe
-    std::vector<std::shared_ptr<Subscriber::PublishNamespaceHandle>> publishNamespaceHandles;
-    std::vector<std::shared_ptr<Publisher::SubscribeNamespaceHandle>> subscribeNamespaceHandles;
-    std::vector<std::shared_ptr<Publisher::SubscriptionHandle>> subscribeHandles;
-
-    void cleanup() {
-      // Simulate MoQSession::cleanup() for publish tracks
-      // This calls publishDone on all tracked consumers, which triggers
-      // FilterConsumer callbacks that properly clean up relay state
-      for (auto& consumer : publishConsumers) {
-        consumer->publishDone(
-            {RequestID(0), PublishDoneStatusCode::SESSION_CLOSED, 0, "mock session cleanup"}
-        );
-      }
-      publishConsumers.clear();
-
-      // Clean up publishNamespaceHandles
-      for (auto& handle : publishNamespaceHandles) {
-        if (handle) {
-          handle->publishNamespaceDone();
-        }
-      }
-      publishNamespaceHandles.clear();
-
-      // Clean up subscribeNamespaceHandles
-      for (auto& handle : subscribeNamespaceHandles) {
-        if (handle) {
-          handle->unsubscribeNamespace();
-        }
-      }
-      subscribeNamespaceHandles.clear();
-
-      // Clean up subscribeHandles
-      for (auto& handle : subscribeHandles) {
-        if (handle) {
-          handle->unsubscribe();
-        }
-      }
-      subscribeHandles.clear();
-    }
-  };
-
-  std::map<MoQSession*, std::shared_ptr<MockSessionState>> mockSessions_;
-
-  std::shared_ptr<MockSessionState> getOrCreateMockState(std::shared_ptr<MoQSession> session) {
-    auto it = mockSessions_.find(session.get());
-    if (it == mockSessions_.end()) {
-      auto state = std::make_shared<MockSessionState>();
-      state->session = session;
-      mockSessions_[session.get()] = state;
-      return state;
-    }
-    return it->second;
-  }
-
-  void cleanupMockSession(std::shared_ptr<MoQSession> session) {
-    auto it = mockSessions_.find(session.get());
-    if (it != mockSessions_.end()) {
-      // Use withSessionContext to ensure session context is set during cleanup
-      withSessionContext(it->second->session, [&]() { it->second->cleanup(); });
-      mockSessions_.erase(it);
-    }
-  }
-
-  // Helper to publish a namespace
-  // Returns the PublishNamespaceHandle so tests can use it for manual cleanup
-  // if needed If addToState is true, the handle is automatically saved for
-  // cleanup
-  std::shared_ptr<Subscriber::PublishNamespaceHandle> doPublishNamespace(
-      std::shared_ptr<MoQSession> session,
-      const TrackNamespace& ns,
-      bool addToState = true
-  ) {
-    PublishNamespace ann;
-    ann.trackNamespace = ns;
-    return withSessionContext(session, [&]() {
-      auto task = relay_->publishNamespace(std::move(ann), nullptr);
-      auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-      EXPECT_TRUE(res.hasValue());
-      if (res.hasValue()) {
-        if (addToState) {
-          getOrCreateMockState(session)->publishNamespaceHandles.push_back(*res);
-        }
-        return *res;
-      }
-      return std::shared_ptr<Subscriber::PublishNamespaceHandle>(nullptr);
-    });
-  }
-
-  // Helper to publish a track
-  // Returns the TrackConsumer so tests can use it for manual operations if
-  // needed If addToState is true, the consumer is automatically saved for
-  // cleanup
-  std::shared_ptr<TrackConsumer> doPublish(
-      std::shared_ptr<MoQSession> session,
-      const FullTrackName& trackName,
-      bool addToState = true
-  ) {
-    PublishRequest pub;
-    pub.fullTrackName = trackName;
-    return withSessionContext(session, [&]() {
-      auto res = relay_->publish(std::move(pub), createMockSubscriptionHandle());
-      EXPECT_TRUE(res.hasValue());
-      if (res.hasValue()) {
-        if (addToState) {
-          getOrCreateMockState(session)->publishConsumers.push_back(res->consumer);
-        }
-        return res->consumer;
-      }
-      return std::shared_ptr<TrackConsumer>(nullptr);
-    });
-  }
-
-  // Helper to subscribe to namespace publishes
-  // Returns the SubscribeNamespaceHandle so tests can use it for manual cleanup
-  // if needed If addToState is true, the handle is automatically saved for
-  // cleanup
-  std::shared_ptr<Publisher::SubscribeNamespaceHandle> doSubscribeNamespace(
-      std::shared_ptr<MoQSession> session,
-      const TrackNamespace& nsPrefix,
-      bool addToState = true
-  ) {
-    SubscribeNamespace subNs;
-    subNs.trackNamespacePrefix = nsPrefix;
-    return withSessionContext(session, [&]() {
-      auto task = relay_->subscribeNamespace(std::move(subNs), nullptr);
-      auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-      EXPECT_TRUE(res.hasValue());
-      if (res.hasValue()) {
-        if (addToState) {
-          getOrCreateMockState(session)->subscribeNamespaceHandles.push_back(*res);
-        }
-        return *res;
-      }
-      return std::shared_ptr<Publisher::SubscribeNamespaceHandle>(nullptr);
-    });
-  }
-
-  // Helper to set up a mock subgroup consumer with default expectations
-  std::shared_ptr<MockSubgroupConsumer> createMockSubgroupConsumer() {
-    auto sg = std::make_shared<NiceMock<MockSubgroupConsumer>>();
-    ON_CALL(*sg, beginObject(_, _, _, _))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-    ON_CALL(*sg, objectPayload(_, _))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(ObjectPublishStatus::IN_PROGRESS)
-        ));
-    ON_CALL(*sg, endOfSubgroup())
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-    ON_CALL(*sg, endOfGroup(_))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-    ON_CALL(*sg, endOfTrackAndGroup(_))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-    return sg;
-  }
-
-  // Publish a track with a caller-supplied handle so tests can set expectations
-  // on requestUpdateCalled.
-  std::shared_ptr<TrackConsumer> doPublishWithHandle(
-      std::shared_ptr<MoQSession> session,
-      const FullTrackName& trackName,
-      std::shared_ptr<Publisher::SubscriptionHandle> handle
-  ) {
-    return withSessionContext(session, [&]() -> std::shared_ptr<TrackConsumer> {
-      PublishRequest pub;
-      pub.fullTrackName = trackName;
-      auto res = relay_->publish(std::move(pub), std::move(handle));
-      EXPECT_TRUE(res.hasValue());
-      if (!res.hasValue()) {
-        return nullptr;
-      }
-      auto consumer = res->consumer;
-      getOrCreateMockState(session)->publishConsumers.push_back(consumer);
-      return consumer;
-    });
-  }
-
-  // Subscribe to a namespace with an explicit forward flag.
-  std::shared_ptr<Publisher::SubscribeNamespaceHandle> doSubscribeNamespaceWithForward(
-      std::shared_ptr<MoQSession> session,
-      const TrackNamespace& nsPrefix,
-      bool forward
-  ) {
-    SubscribeNamespace subNs;
-    subNs.trackNamespacePrefix = nsPrefix;
-    subNs.forward = forward;
-    return withSessionContext(session, [&]() {
-      auto task = relay_->subscribeNamespace(std::move(subNs), nullptr);
-      auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-      EXPECT_TRUE(res.hasValue());
-      if (!res.hasValue()) {
-        return std::shared_ptr<Publisher::SubscribeNamespaceHandle>(nullptr);
-      }
-      getOrCreateMockState(session)->subscribeNamespaceHandles.push_back(*res);
-      return *res;
-    });
-  }
-
-  // Set up session->publish() to succeed with a mock consumer and immediate
-  // PublishOk. Call this on any subscriber session before it will receive
-  // publishToSession() calls, otherwise the mock returns an error and the
-  // subscriber is ejected, confusing REQUEST_UPDATE accounting.
-  void setupPublishSucceeds(std::shared_ptr<MockMoQSession> session) {
-    ON_CALL(*session, publish(_, _))
-        .WillByDefault(Invoke([this](PublishRequest pub, auto) -> Subscriber::PublishResult {
-          PublishOk ok{
-              pub.requestID,
-              /*forward=*/pub.forward,
-              /*priority=*/128,
-              GroupOrder::Default,
-              LocationType::LargestObject,
-              /*start=*/std::nullopt,
-              /*endGroup=*/std::make_optional(uint64_t(0))
-          };
-          return Subscriber::PublishConsumerAndReplyTask{
-              createMockConsumer(),
-              folly::coro::makeTask<folly::Expected<PublishOk, PublishError>>(std::move(ok))
-          };
-        }));
-  }
-
-  // Build a NiceMock SubscriptionHandle suitable for publish() calls.
-  std::shared_ptr<NiceMock<MockSubscriptionHandle>> makePublishHandle() {
-    SubscribeOk ok;
-    ok.requestID = RequestID(0);
-    ok.trackAlias = TrackAlias(0);
-    ok.expires = std::chrono::milliseconds(0);
-    ok.groupOrder = GroupOrder::Default;
-    auto handle = std::make_shared<NiceMock<MockSubscriptionHandle>>(std::move(ok));
-    ON_CALL(*handle, requestUpdateResult())
-        .WillByDefault(Return(folly::makeExpected<RequestError>(RequestOk{})));
-    return handle;
-  }
-
-  std::shared_ptr<TestMoQExecutor> exec_;
-  std::shared_ptr<MoqxRelay> relay_;
-};
 
 // Test: Basic relay construction
 TEST_F(MoQRelayTest, Construction) {
@@ -433,6 +48,9 @@ TEST_F(MoQRelayTest, PublishSuccess) {
   // Cleanup: remove the session from relay to avoid mock leak warning
   removeSession(publisherSession);
 }
+
+// Namespace tree tests (Prune*, MixedContent*, ActiveChildCount, PublishKeepsNode)
+// are in NamespaceTreeTest.cpp.
 
 // Test: Verify that subgroups are only created at appropriate times
 // Sequence: publish, sub1 joins, beginSubgroup (->1), beginObject (->1),
@@ -3260,6 +2878,7 @@ TEST_F(MoQRelayTest, PublishReconnectDuringSubscribeSuccessPathCrash) {
   relay_->doPublishNamespaceDone(kTestNamespace, publisherSession2);
 }
 
+// Regression: after publishDone the namespace-tree node must be pruned when
 // the track was the only remaining content.  (Was a bug before unpublishTrack
 // gained a NodeMutationGuard; kept as a regression guard.)
 TEST_F(MoQRelayTest, PublishDonePrunesNamespaceTreeNode) {
